move errors to root of ddbretry package
