+++ ddbretry/ddbretry.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"time"
 
-	"github.com/Thumbscrew/aws-go-tools/ddbretry/errors"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb/types"
 )
@@ -50,7 +49,7 @@ func (c *RetryDynamoDBClient) DeleteItem(ctx context.Context, input *dynamodb.De
 		}
 	}
 
-	return nil, errors.NewInvalidRetryError(retries)
+	return nil, NewInvalidRetryError(retries)
 }
 
 func IsProvisionedThroughputExceededException(err error) bool {
